papa relaciones con cajones

+++ src/main/java/com/example/demo/repository/PapaRepository.java
@@ -4,6 +4,6 @@ import com.example.demo.domain.Papa;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface PapaRepository extends JpaRepository<Papa, Integer> {
+public interface PapaRepository extends JpaRepository<Papa, Long> {
 
 }
